MODEXPW-545 - Improve caching
