+++ src/main/java/org/folio/dew/service/HoldingsReferenceService.java
@@ -6,27 +6,17 @@ import static org.apache.commons.lang3.StringUtils.EMPTY;
 import static org.apache.commons.lang3.StringUtils.isEmpty;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-import org.apache.commons.lang3.ObjectUtils;
-import org.folio.dew.client.CallNumberTypeClient;
 import org.folio.dew.client.HoldingClient;
-import org.folio.dew.client.HoldingsNoteTypeClient;
-import org.folio.dew.client.HoldingsSourceClient;
-import org.folio.dew.client.HoldingsTypeClient;
-import org.folio.dew.client.IllPolicyClient;
 import org.folio.dew.client.InstanceClient;
 import org.folio.dew.client.InventoryClient;
-import org.folio.dew.client.LocationClient;
-import org.folio.dew.client.StatisticalCodeClient;
 import org.folio.dew.domain.dto.ErrorServiceArgs;
 import org.folio.dew.domain.dto.HoldingsRecord;
 import org.folio.dew.error.BulkEditException;
 import org.folio.dew.error.NotFoundException;
 import org.folio.spring.FolioExecutionContext;
 import org.folio.spring.scope.FolioExecutionContextSetter;
-import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
 
@@ -39,16 +29,10 @@ public class HoldingsReferenceService extends FolioExecutionContextManager {
 
   private final InstanceClient instanceClient;
   private final InventoryClient inventoryClient;
-  private final HoldingsTypeClient holdingsTypeClient;
-  private final LocationClient locationClient;
-  private final CallNumberTypeClient callNumberTypeClient;
-  private final HoldingsNoteTypeClient holdingsNoteTypeClient;
-  private final IllPolicyClient illPolicyClient;
-  private final HoldingsSourceClient sourceClient;
-  private final StatisticalCodeClient statisticalCodeClient;
   private final BulkEditProcessingErrorsService errorsService;
   private final HoldingClient holdingClient;
   private final FolioExecutionContext folioExecutionContext;
+  private final HoldingsReferenceServiceCache holdingsReferenceServiceCache;
 
   public HoldingsRecord getHoldingById(String id, String tenantId) {
     try (var context = new FolioExecutionContextSetter(refreshAndGetFolioExecutionContext(tenantId, folioExecutionContext))) {
@@ -109,19 +93,9 @@ public class HoldingsReferenceService extends FolioExecutionContextManager {
     return items.getItems().get(0).getHoldingsRecordId();
   }
 
-  @Cacheable(cacheNames = "holdingsTypesNames")
-  public String getHoldingsTypeNameById(String id, String tenantId) {
-   if (isEmpty(id)) {
-      return EMPTY;
-    }
-    try (var context = new FolioExecutionContextSetter(refreshAndGetFolioExecutionContext(tenantId, folioExecutionContext))) {
-      return holdingsTypeClient.getById(id).getName();
-    }
-  }
-
   public String getHoldingsTypeNameById(String id, ErrorServiceArgs args, String tenantId) {
     try  {
-      return getHoldingsTypeNameById(id, tenantId);
+      return holdingsReferenceServiceCache.getHoldingsTypeNameById(id, tenantId);
     } catch (NotFoundException e) {
       var msg = "Holdings type not found by id=" + id;
       log.error(msg);
@@ -130,30 +104,9 @@ public class HoldingsReferenceService extends FolioExecutionContextManager {
     }
   }
 
-  @Cacheable(cacheNames = "holdingsLocationsNames")
-  public String getLocationNameById(String id, String tenantId) {
-    try (var context = new FolioExecutionContextSetter(refreshAndGetFolioExecutionContext(tenantId, folioExecutionContext))) {
-      return isEmpty(id) ? EMPTY : locationClient.getLocationById(id).getName();
-    } catch (NotFoundException e) {
-      var msg = "Location not found by id=" + id;
-      log.error(msg);
-      throw new BulkEditException(msg);
-    }
-  }
-
-  @Cacheable(cacheNames = "holdingsCallNumberTypesNames")
-  public String getCallNumberTypeNameById(String id, String tenantId) {
-    if (isEmpty(id)) {
-      return EMPTY;
-    }
-    try (var context = new FolioExecutionContextSetter(refreshAndGetFolioExecutionContext(tenantId, folioExecutionContext))) {
-      return callNumberTypeClient.getById(id).getName();
-    }
-  }
-
   public String getCallNumberTypeNameById(String id, ErrorServiceArgs args, String tenantId) {
     try  {
-      return getCallNumberTypeNameById(id, tenantId);
+      return holdingsReferenceServiceCache.getCallNumberTypeNameById(id, tenantId);
     } catch (NotFoundException e) {
       var msg = "Call number type not found by id=" + id;
       log.error(msg);
@@ -162,19 +115,9 @@ public class HoldingsReferenceService extends FolioExecutionContextManager {
     }
   }
 
-  @Cacheable(cacheNames = "holdingsNoteTypesNames")
-  public String getNoteTypeNameById(String id, String tenantId) {
-    if (isEmpty(id)) {
-      return EMPTY;
-    }
-    try (var context = new FolioExecutionContextSetter(refreshAndGetFolioExecutionContext(tenantId, folioExecutionContext))) {
-      return holdingsNoteTypeClient.getById(id).getName();
-    }
-  }
-
   public String getNoteTypeNameById(String id, ErrorServiceArgs args, String tenantId) {
     try {
-      return getNoteTypeNameById(id, tenantId);
+      return holdingsReferenceServiceCache.getNoteTypeNameById(id, tenantId);
     } catch (NotFoundException e) {
       var msg = "Note type not found by id=" + id;
       log.error(msg);
@@ -183,19 +126,9 @@ public class HoldingsReferenceService extends FolioExecutionContextManager {
     }
   }
 
-  @Cacheable(cacheNames = "illPolicyNames")
-  public String getIllPolicyNameById(String id, String tenantId) {
-    if (isEmpty(id)) {
-      return EMPTY;
-    }
-    try (var context = new FolioExecutionContextSetter(refreshAndGetFolioExecutionContext(tenantId, folioExecutionContext))) {
-      return illPolicyClient.getById(id).getName();
-    }
-  }
-
   public String getIllPolicyNameById(String id, ErrorServiceArgs args, String tenantId) {
     try {
-      return getIllPolicyNameById(id, tenantId);
+      return holdingsReferenceServiceCache.getIllPolicyNameById(id, tenantId);
     } catch (NotFoundException e) {
       var msg = "Ill policy not found by id=" + id;
       log.error(msg);
@@ -204,19 +137,10 @@ public class HoldingsReferenceService extends FolioExecutionContextManager {
     }
   }
 
-  @Cacheable(cacheNames = "holdingsSourceNames")
-  public String getSourceNameById(String id, String tenantId) {
-    if (isEmpty(id)) {
-      return EMPTY;
-    }
-    try (var context = new FolioExecutionContextSetter(refreshAndGetFolioExecutionContext(tenantId, folioExecutionContext))) {
-      return sourceClient.getById(id).getName();
-    }
-  }
 
   public String getSourceNameById(String id, ErrorServiceArgs args, String tenantId) {
     try {
-      return getSourceNameById(id, tenantId);
+      return holdingsReferenceServiceCache.getSourceNameById(id, tenantId);
     } catch (NotFoundException e) {
       var msg = "Holdings record source not found by id=" + id;
       log.error(msg);
@@ -225,19 +149,9 @@ public class HoldingsReferenceService extends FolioExecutionContextManager {
     }
   }
 
-  @Cacheable(cacheNames = "holdingsStatisticalCodeNames")
-  public String getStatisticalCodeNameById(String id, String tenantId) {
-    if (isEmpty(id)) {
-      return EMPTY;
-    }
-    try (var context = new FolioExecutionContextSetter(refreshAndGetFolioExecutionContext(tenantId, folioExecutionContext))) {
-      return statisticalCodeClient.getById(id).getName();
-    }
-  }
-
   public String getStatisticalCodeNameById(String id, ErrorServiceArgs args, String tenantId) {
     try {
-      return getStatisticalCodeNameById(id, tenantId);
+      return holdingsReferenceServiceCache.getStatisticalCodeNameById(id, tenantId);
     } catch (NotFoundException e) {
       var msg = "Statistical code not found by id=" + id;
       log.error(msg);
@@ -246,20 +160,15 @@ public class HoldingsReferenceService extends FolioExecutionContextManager {
     }
   }
 
-  @Cacheable(cacheNames = "holdings")
   public JsonNode getHoldingsJsonById(String holdingsId, String tenantId) {
-    try (var context = new FolioExecutionContextSetter(refreshAndGetFolioExecutionContext(tenantId, folioExecutionContext))) {
-      return holdingClient.getHoldingById(holdingsId);
-    }
+    return  holdingsReferenceServiceCache.getHoldingsJsonById(holdingsId, tenantId);
   }
 
-  @Cacheable(cacheNames = "holdingsLocations")
   public JsonNode getHoldingsLocationById(String locationId, String tenantId) {
-    if (ObjectUtils.isEmpty(locationId)) {
-      return new ObjectMapper().createObjectNode();
-    }
-    try (var context = new FolioExecutionContextSetter(refreshAndGetFolioExecutionContext(tenantId, folioExecutionContext))) {
-      return locationClient.getLocation(locationId);
-    }
+   return  holdingsReferenceServiceCache.getHoldingsLocationById(locationId, tenantId);
+  }
+
+  public String getLocationNameById(String id, String tenantId) {
+   return holdingsReferenceServiceCache.getLocationNameById(id, tenantId);
   }
 }
